+++ src/deluge/model/clip/audio_clip.cpp
@@ -236,7 +236,7 @@ void AudioClip::finishLinearRecording(ModelStackWithTimelineCounter* modelStack,
 
 	recorder = NULL;
 
-	clipName.set(sampleHolder.filePath.get());
+	name.set(sampleHolder.filePath.get());
 }
 
 Clip* AudioClip::cloneAsNewOverdub(ModelStackWithTimelineCounter* modelStackOldClip, OverDubType newOverdubNature) {
@@ -1212,7 +1212,7 @@ Error AudioClip::claimOutput(ModelStackWithTimelineCounter* modelStack) {
 
 void AudioClip::loadSample(bool mayActuallyReadFile) {
 	Error error = sampleHolder.loadFile(sampleControls.reversed, false, mayActuallyReadFile);
-	clipName.set(sampleHolder.filePath.get());
+	name.set(sampleHolder.filePath.get());
 	if (error != Error::NONE) {
 		display->displayError(error);
 	}
@@ -1290,7 +1290,7 @@ void AudioClip::clear(Action* action, ModelStackWithTimelineCounter* modelStack,
 
 			sampleHolder.filePath.clear();
 			sampleHolder.setAudioFile(NULL);
-			clipName.set("");
+			name.set("");
 		}
 
 		renderData.xScroll = -1;