+++ src/deluge/model/clip/instrument_clip_minder.cpp
@@ -48,6 +48,7 @@
 #include "processing/sound/sound_instrument.h"
 #include "util/lookuptables/lookuptables.h"
 #include <cstring>
+#include <string.h>
 
 extern "C" {
 #include "RZA1/uart/sio_char.h"
@@ -115,42 +116,71 @@ void InstrumentClipMinder::renderOLED(deluge::hid::display::oled_canvas::Canvas&
 #pragma GCC diagnostic ignored "-Wstack-usage="
 void InstrumentClipMinder::drawMIDIControlNumber(int32_t controlNumber, bool automationExists) {
 
-	char buffer[display->haveOLED() ? 30 : 5];
-	bool finish = false;
+	DEF_STACK_STRING_BUF(buffer, 30);
+
+	bool doScroll = false;
+
 	if (controlNumber == CC_NUMBER_NONE) {
-		strcpy(buffer, deluge::l10n::get(deluge::l10n::String::STRING_FOR_NO_PARAM));
+		buffer.append(deluge::l10n::get(deluge::l10n::String::STRING_FOR_NO_PARAM));
 	}
 	else if (controlNumber == CC_NUMBER_PITCH_BEND) {
-		strcpy(buffer, deluge::l10n::get(deluge::l10n::String::STRING_FOR_PITCH_BEND));
+		buffer.append(deluge::l10n::get(deluge::l10n::String::STRING_FOR_PITCH_BEND));
 	}
 	else if (controlNumber == CC_NUMBER_AFTERTOUCH) {
-		strcpy(buffer, deluge::l10n::get(deluge::l10n::String::STRING_FOR_CHANNEL_PRESSURE));
+		buffer.append(deluge::l10n::get(deluge::l10n::String::STRING_FOR_CHANNEL_PRESSURE));
 	}
 	else if (controlNumber == CC_NUMBER_Y_AXIS) {
 		// in mono expression this is mod wheel, and y-axis is not directly controllable
-		strcpy(buffer, deluge::l10n::get(deluge::l10n::String::STRING_FOR_MOD_WHEEL));
+		buffer.append(deluge::l10n::get(deluge::l10n::String::STRING_FOR_MOD_WHEEL));
 	}
 	else {
-		buffer[0] = 'C';
-		buffer[1] = 'C';
-		if (display->haveOLED()) {
-			buffer[2] = ' ';
-			intToString(controlNumber, &buffer[3]);
+		MIDIInstrument* midiInstrument = (MIDIInstrument*)getCurrentOutput();
+
+		String name;
+
+		if (controlNumber >= 0 && controlNumber < kNumRealCCNumbers) {
+			midiInstrument->getNameFromCC(controlNumber, &name);
 		}
 		else {
-			char* numberStartPos = (controlNumber < 100) ? (buffer + 2) : (buffer + 1);
-			intToString(controlNumber, numberStartPos);
+			name.clear();
+		}
+
+		// if we have a name for this midi cc set by the user, display that instead of the cc number
+		if (!name.isEmpty()) {
+			buffer.append(name.get());
+			doScroll = name.getLength() > 4 ? true : false;
+		}
+		else {
+			if (display->haveOLED()) {
+				buffer.append("CC ");
+				buffer.appendInt(controlNumber);
+			}
+			else {
+				if (controlNumber < 100) {
+					buffer.append("CC");
+				}
+				else {
+					buffer.append("C");
+				}
+				buffer.appendInt(controlNumber);
+			}
 		}
 	}
 
 	if (display->haveOLED()) {
 		if (automationExists) {
-			strcat(buffer, "\n(automated)");
+			buffer.append("\n");
+			buffer.append("(automated)");
 		}
-		display->popupText(buffer);
+		display->popupText(buffer.c_str());
 	}
 	else {
-		display->setText(buffer, true, automationExists ? 3 : 255, false);
+		if (doScroll) {
+			display->setScrollingText(buffer.c_str(), 0, 600, -1, automationExists ? 3 : 255);
+		}
+		else {
+			display->setText(buffer.c_str(), true, automationExists ? 3 : 255, false);
+		}
 	}
 }
 #pragma GCC pop